create landing to championship

+++ src/app/core/interceptors/auth.interceptor.ts
@@ -1,68 +1,69 @@
-import { HttpRequest, HttpHandlerFn, HttpEvent, HttpErrorResponse } from '@angular/common/http';
-import { Observable, throwError } from 'rxjs';
-import { catchError, switchMap } from 'rxjs/operators';
-import { inject } from '@angular/core';
-import { AuthService } from '../services/auth.service';
-import { Router } from '@angular/router';
-import { LoadingService } from '../services/loading.service';
+import { HttpRequest, HttpHandlerFn, HttpEvent, HttpErrorResponse } from '@angular/common/http'
+import { Observable, throwError } from 'rxjs'
+import { catchError, switchMap } from 'rxjs/operators'
+import { inject } from '@angular/core'
+import { AuthService } from '../services/auth.service'
+import { Router } from '@angular/router'
+import { LoadingService } from '../services/loading.service'
 
 export function authInterceptorFn(req: HttpRequest<unknown>, next: HttpHandlerFn): Observable<HttpEvent<unknown>> {
-  const authService = inject(AuthService);
-  const router = inject(Router);
-  const token = authService.getToken();
-  const loadingService = inject(LoadingService);
+  const authService = inject(AuthService)
+  const router = inject(Router)
+  const token = authService.getToken()
+  const loadingService = inject(LoadingService)
 
-  let authReq = req;
+  let authReq = req
+  if (req.url.includes('/authorize/update')) {
+    return next(req);
+  }
   if (token) {
     authReq = req.clone({
       setHeaders: {
         Authorization: `Bearer ${token}`,
       },
-    });
+    })
   }
 
   return next(authReq).pipe(
     catchError((error: HttpErrorResponse) => {
+      if (req.url.includes('/authorize/update')) {
+        return throwError(() => error)
+      }
       if (error.status === 401) {
-        const refresh = authService.getRefreshToken();
+        const refresh = authService.getRefreshToken()
         if (!refresh) {
-          // loadingService.show(); // muestra loading
-          // setTimeout(() => {
-            authService.removeToken();
-            authService.removeRefreshToken()
-
-            // window.location.reload(); // recarga la página completamente
-          // }, 2000); // breve delay para mostrar spinner
-          return throwError(() => error);
+          authService.removeToken()
+          authService.removeRefreshToken()
+          return throwError(() => error)
         }
 
         return authService.refreshToken().pipe(
           switchMap((res: any) => {
-            authService.saveToken(res.access_token);
-            authService.saveRefreshToken(res.refresh_token);
+            authService.saveToken(res.access_token)
+            authService.saveRefreshToken(res.refresh_token)
             const newAuthReq = req.clone({
               setHeaders: {
                 Authorization: `Bearer ${res.access_token}`,
               },
-            });
-
-            return next(newAuthReq); // Reintenta la solicitud original con nuevo token
+            })
+            return next(newAuthReq)
           }),
           catchError(err => {
-            loadingService.show();
             setTimeout(() => {
-              authService.removeToken();
-              authService.removeRefreshToken();
-              window.location.reload();
-            }, 500);
-            return throwError(() => err);
-          })
-        );
+              authService.removeToken()
+              authService.removeRefreshToken()
+              // authService.removeUser()
+              router.navigate(['/login'])
+              window.location.reload()
+            }, 500)
+            return throwError(() => err)
+          }),
+        )
       } else if (error.status === 403) {
-        router.navigate(['/login']);
+        router.navigate(['/login'])
       }
 
-      return throwError(() => error);
-    })
-  );
+      return throwError(() => error)
+    }),
+  )
 }
